finish to do

+++ include/Util.h
@@ -31,7 +31,7 @@ typedef struct State {
 State_t* new_State();
 void print_prompt(State_t *state);
 void print_user(User_t *user, SelectArgs_t *sel_args);
-void print_users(Table_t *table, int *idxList, size_t idxListLen, Command_t *cmd);
+void print_users(Table_t *table, size_t *idxList, size_t idxListLen, Command_t *cmd);
 int parse_input(char *input, Command_t *cmd);
 void handle_builtin_cmd(Table_t *table, Command_t *cmd, State_t *state);
 int handle_query_cmd(Table_t *table, Command_t *cmd);
@@ -41,5 +41,5 @@ int handle_update_cmd(Table_t *table, Command_t *cmd);
 int handle_delete_cmd(Table_t *table, Command_t *cmd);
 int handle_join_cmd(Table_t *table1, Table_Like_t *table2, Command_t *cmd);
 void print_help_msg();
-size_t where_check(Table_t *table, Command_t *cmd, int* idxList, Hash_t* hash_entry);
+size_t where_check(Table_t *table, Command_t *cmd, size_t* idxList, Hash_t* hash_entry);
 #endif
